+++ iSkyLIMS_wetlab/views.py
@@ -34,11 +34,9 @@ from .utils.library_preparation import *
 from .utils.pool_preparation import *
 from .utils.run_preparation import *
 from .utils.additional_kits import *
-from .utils.handling_statistics import *
-from .utils.handling_sequencers import *
+from .utils.statistics import *
+from .utils.sequencers import *
 
-# from .utils.samplesheet_checks import *
-# from .utils.wetlab_misc_utilities import normalized_data
 from iSkyLIMS_core.utils.handling_samples import *
 from iSkyLIMS_core.utils.handling_platforms import get_defined_platforms_and_ids
 from iSkyLIMS_core.utils.common import (
@@ -486,34 +484,7 @@ def create_nextseq_run(request):
                 library[project_index_kit[x]].append(projects[x])
             else:
                 library[project_index_kit[x]] = [projects[x]]
-        ## convert the sample sheet to base space format and have different files according the library kit
-        """ Removed because Base space creation file is not longer used
-        for key, value in library.items():
-            lib_kit_file =key.replace(' ', '_')
-            library_file = sample_sheet_map_basespace(in_file, key, lib_kit_file, value,'Plate96')
-            if library_file == 'ERROR':
-                # deleting the sample sheet file
-                os.remove(in_file)
-                # Deleting projects related to the shample sheet
-                for p in range(len( projects)):
-                    my_project = projects [p]
-                    delete_proj=Projects.objects.get(projectName = my_project)
-                    delete_proj.delete()
-                # delete the run used when uploading the sample sheet
-                run_p.delete()
-                # show the error page
-                return render (
-                    request,'iSkyLIMS_wetlab/error_page.html',
-                    {'content':[ 'The information on  the Library kit ', key,
-                    ' For the project ', value,
-                    'Does not meet the requirements to perform the conversion to import to Base Space',
-                    'ADVICE', 'Check the sample sheet that was uploaded ']})
-            else:
-                bs_file[key] = library_file
-                results.append([key, bs_file[key]])
-        """
         ## save the project information on database
-
         for p in range(len(projects)):
             my_project = projects[p]
             my_name = user_name[p]
@@ -530,32 +501,6 @@ def create_nextseq_run(request):
             update_info_proj.user_id = User.objects.get(username__exact=user_name[p])
             update_info_proj.save()
         results.append(["runname", experiment_name])
-        ## save the sample sheet file under tmp/recorded to be processed when run folder was created
-        """
-        ########
-        ## Remove the folder creation to include the sample sheet from iSkyLIMS version 2
-        subfolder_name=str(run_p.id)
-
-        temp_directory = os.path.join(settings.MEDIA_ROOT , wetlab_config.RUN_TEMP_DIRECTORY_RECORDED, subfolder_name)
-        if not os.path.isdir(temp_directory):
-            os.makedirs(temp_directory)
-        # os.mkdir(temp_directory)
-        # set group writing permission to the temporary directory
-        os.chmod(temp_directory, 0o774)
-        #os.mkdir(os.path.join(settings.MEDIA_ROOT, 'wetlab/tmp/recorded', subfolder_name ))
-        sample_sheet_copy= os.path.join(temp_directory, 'samplesheet.csv' )
-        shutil.copy(in_file,sample_sheet_copy)
-        # set the group write permission to the Sample Sheet File
-        os.chmod(sample_sheet_copy, 0o664)
-        # update the sample sheet with the experiment name
-        if run_name != experiment_name :
-            update_sample_sheet (in_file, experiment_name)
-        ## update the Experiment name and the state of the run to 'Recorded'
-        run_p.runName = experiment_name
-        run_p.index_library = run_index_library_name
-        run_p.save()
-        #######################
-        """
         run_p.set_run_state("Recorded")
         sample_sheet_lines = read_all_lines_in_sample_sheet(in_file)
         sample_names_and_data = get_samples_in_sample_sheet(sample_sheet_lines)
@@ -1930,133 +1875,6 @@ def change_project_libKit(request, project_id):
         )
 
 
-"""
-@login_required
-def change_run_libKit (request, run_id):
-    #check if run exist
-    if RunProcess.objects.filter(pk = run_id).exists():
-        run_obj = RunProcess.objects.get(pk = run_id)
-        if not request.user.is_authenticated :
-            return redirect ('/accounts/login')
-        # check if user is allow to make the change
-        groups = Group.objects.get(name='WetlabManager')
-        # check if user belongs to WetlabManager . If true allow to see the page
-        if groups not in request.user.groups.all():
-            return render (request,'iSkyLIMS_wetlab/error_page.html', {'content':['You do have the enough privileges to see this page ','Contact with your administrator .']})
-        if request.method == 'POST' and request.POST['action'] == 'change_run_libKit':
-            new_library_kit = request.POST.getlist('runlibraryKit')
-            projects_name = request.POST.getlist('projectInRun')
-            changed_lib_kit_dict = {}
-
-            # check if there is only one library kit associated to the run
-            if len(new_library_kit) == 1 :
-                # Check if new library kit was set in the form
-                project = Projects.objects.get(projectName__exact = projects_name[0])
-                old_library_kit = project.get_index_library_name()
-                if new_library_kit[0] == old_library_kit :
-                    return render (request, 'iSkyLIMS_wetlab/info_page.html', {'content': ['The library kit from the input text is the same to the existing defined for this project', 'No change is done']})
-                # change the library name
-                old_lib_kit_file = project.baseSpaceFile
-                new_file_name = new_library_kit[0].replace(' ' , '_')
-                #
-
-                new_file = update_library_kit_field(old_lib_kit_file,new_file_name,new_library_kit[0])
-                if new_file == 'ERROR':
-                    return render (request, 'iSkyLIMS_wetlab/error_page.html', {'content':['']})
-                # update the database with new file
-                project.baseSpaceFile = new_file
-                project.libraryKit = new_library_kit[0]
-                project.save()
-
-            else:
-                old_files_to_be_deleted = []
-                # check if any of the library has change
-                need_to_be_updated = False
-                for item in range(len(projects_name)):
-                    project = Projects.objects.get(projectName__exact = projects_name[item])
-                    old_library_kit = project.libraryKit
-                    # get the library kit file name to delete it later
-                    old_file = project.baseSpaceFile
-                    # build the list to delete later the old library kit files
-                    if old_file not in old_files_to_be_deleted :
-                        old_files_to_be_deleted.append(old_file)
-                    if new_library_kit[item] != old_library_kit:
-                        need_to_be_updated = True
-                if not need_to_be_updated:
-                    return render (request, 'iSkyLIMS_wetlab/info_page.html', {'content': ['The library kits from the input text are the same to the existing ones defined for the Run', 'No change is done']})
-                # get the Sample Sheet file related to the run
-                sample_file = run.get_sample_file()
-
-                lib_kit_dict = {}
-                in_file=str('documents/' + sample_file)
-                #
-                ## build the project list for each library kit
-                for x in range(len(new_library_kit)):
-                    if new_library_kit[x] in lib_kit_dict :
-                        lib_kit_dict[new_library_kit[x]].append(projects_name[x])
-                    else:
-                        lib_kit_dict[new_library_kit[x]]= [projects_name[x]]
-
-                ## convert the sample sheet to base space format and have different files according the library kit
-                #
-
-                for key, value in lib_kit_dict.items():
-                    lib_kit_file =key.replace(' ', '_')
-                    library_file = sample_sheet_map_basespace(in_file, key, lib_kit_file, value,'Plate96')
-                    if library_file == 'ERROR':
-                        return render (request,'iSkyLIMS_wetlab/error_page.html', {'content':[ 'The information on  the Library kit ', key,' For the project ', value,
-                          'could not be changed to the new value of the library kit ','ADVICE', 'Contact your administrator']})
-                    for updated_project_name in value :
-                        p_updated = Projects.objects.get(projectName__exact = updated_project_name)
-                        p_updated.libraryKit = key
-                        p_updated.baseSpaceFile = library_file
-                        p_updated.save()
-                # delete old library kit files
-                for delete_file in old_files_to_be_deleted :
-                    os.remove(delete_file)
-                # prepare the information to be displayed
-            changed_lib_kit_dict = {}
-            run_data = {}
-            for item in range(len(projects_name)) :
-                project = Projects.objects.get(projectName__exact = projects_name[item])
-                p_name =project.projectName
-                lib_name = project.libraryKit
-                lib_file = project.baseSpaceFile
-                run_data[p_name] = ([[lib_name, lib_file]])
-
-            changed_lib_kit_dict['run_name'] = run.get_run_name()
-            changed_lib_kit_dict['run_data'] = run_data
-
-            return render (request, 'iSkyLIMS_wetlab/ChangeRunLibraryKit.html',{'changed_lib_kit': changed_lib_kit_dict})
-
-        else:
-            form_change_lib_kit = {}
-            run_library_data = []
-            project_list = []
-            #library_kit_list = []
-            library_kit_dict = {}
-            form_change_lib_kit['run_name'] = run_obj.get_run_name()
-            # get the library Kits used in run
-            run_obj.get_index_library()
-                pass
-            project_list = Projects.objects.filter(runprocess_id = run_id)
-            for project in project_list :
-                lib_kit = project.libraryKit
-                run_library_data.append([project.projectName, lib_kit])
-                if lib_kit in library_kit_dict :
-                    library_kit_dict[project.libraryKit].append(project.projectName)
-                else :
-                    library_kit_dict[project.libraryKit] = [project.projectName]
-
-            form_change_lib_kit['run_data'] = library_kit_dict
-            form_change_lib_kit['run_library_data'] = run_library_data
-
-            return render (request, 'iSkyLIMS_wetlab/ChangeRunLibraryKit.html',{'form_change_lib_kit': form_change_lib_kit})
-    else:
-        return render (request,'iSkyLIMS_wetlab/error_page.html', {'content':['No run has been found for changing the library Kit ' ]})
-"""
-
-
 @login_required
 def stats_experiment(request):
     return render(request, "iSkyLIMS_wetlab/StatsPerExperiment.html", {})
@@ -2106,213 +1924,6 @@ def stats_per_sequencer(request):
             runs_using_sequencer, sequencer, start_date, end_date
         )
 
-        """
-
-
-                    # Get data from researcher projects
-
-
-
-
-                    sequencer_statistics = {}
-
-                    sequencer_statistics ['run_heading'] = ['Run name' , 'Run Date' , 'Sequencer Model' , 'State' , 'Space Img(Mb)' , 'Space Fasta(Mb)' , 'Space Other(Mb)']
-                    runs_data = []
-                    sequencer_run = {}
-                    runs_data_list = []
-                    runs_time_list = []
-                    for run_machine in run_by_machine:
-                        ru_seq_model = run_machine.get_run_used_sequencer()
-                        ru_data=run_machine.get_info_process().split(';')
-                        #runName, state, requested_center, generated_date, rundate,completed_date, bcl2fastq_date, finish_date, useSpaceImgMb, useSpaceFastaMb,useSpaceOtherMb
-                        runs_data_list.append([ru_data[0], ru_data[4] , ru_seq_model , ru_data[3] , ru_data[8] , ru_data[9] , ru_data[10]])
-                        # time the run took from run_date to completed_date
-                        runs_time_list.append([ru_data[0], ru_data[5]-ru_data[4]])
-
-                    sequencer_run[ru_seq_model] = runs_data_list
-                    runs_data.append(sequencer_run)
-                    sequencer_statistics['run_data'] = runs_data
-                    # Get data from machine projects
-
-                    sequencer_statistics ['projects_heading'] = ['Project name', 'Date', 'Libraty Kit','Samples', 'Cluster PF', 'Yield Mb', '% Q> 30', 'Mean','Sequencer ID']
-                    projects_data =[]
-                    p_machine_date , p_machine_num_sample = {} , {}
-                    p_machine_lib_kit, p_machine_sequencer = {}, {}
-                    p_machine_q30_dict, p_machine_mean_dict = {} , {}
-                    p_machine_yield_mb_dict, p_machine_cluster_pf_dict ={} , {}
-                    projects_name_dict , projects_id_list = {} , {}
-
-
-
-                    for project_machine in projects_by_machine:
-
-                        q_30_list , mean_q_list = [] , []
-                        yield_mb_list,  cluster_pf_list = [], []
-                        p_name = project_machine.get_project_name()
-
-                        sequencer_in_project = project_machine.runprocess_id.get_run_used_sequencer()
-                        if not sequencer_in_project in projects_name_dict :
-                            p_machine_num_sample[sequencer_in_project] ={}
-                            p_machine_sequencer[sequencer_in_project] ={}
-                            p_machine_date[sequencer_in_project] ={}
-                            p_machine_lib_kit[sequencer_in_project] ={}
-                            p_machine_q30_dict[sequencer_in_project] ={}
-                            p_machine_mean_dict[sequencer_in_project] ={}
-                            p_machine_yield_mb_dict[sequencer_in_project] ={}
-                            p_machine_cluster_pf_dict[sequencer_in_project] ={}
-                            projects_name_dict[sequencer_in_project] = []
-                            projects_id_list[sequencer_in_project] =  []
-                        projects_name_dict[sequencer_in_project].append(p_name)
-                        m_project_id = project_machine.id
-                        projects_id_list[sequencer_in_project].append(m_project_id)
-                        p_machine_num_sample[sequencer_in_project][p_name] = StatsFlSummary.objects.get(project_id__exact = m_project_id).sampleNumber
-                        p_machine_date [sequencer_in_project][p_name] = project_machine.get_date()
-                        p_machine_lib_kit[sequencer_in_project][p_name]= project_machine.get_index_library_name()
-                        p_machine_sequencer[sequencer_in_project][p_name] = str(project_machine.runprocess_id.sequencerModel)
-                        lanes_in_project = StatsLaneSummary.objects.filter( project_id__exact = m_project_id)
-                        for lane in lanes_in_project :
-                            q_30_value, mean_q_value , yield_mb_value , cluster_pf_value = lane.get_stats_info()
-                            q_30_list.append(float(q_30_value))
-                            mean_q_list.append(float(mean_q_value))
-                            yield_mb_list.append(float(yield_mb_value.replace(',','')))
-                            cluster_pf_list.append(float(cluster_pf_value.replace(',','')))
-                        p_machine_q30_dict[sequencer_in_project] [p_name]= format(statistics.mean(q_30_list), '.2f')
-                        p_machine_mean_dict[sequencer_in_project][p_name] = format(statistics.mean(mean_q_list), '.2f')
-                        p_machine_yield_mb_dict[sequencer_in_project][p_name] = round(sum(yield_mb_list))
-                        p_machine_cluster_pf_dict[sequencer_in_project][p_name] = round(sum(cluster_pf_list))
-
-                    # Create the table with projects executed by the machine
-                    machine_seq_graphs, machine_graphs = [], []
-                    for sequencer, projects_name_list in projects_name_dict.items() :
-                        sequencer_proj = {}
-                        proj_data =[]
-                        for project_name in projects_name_list :
-                            proj_data.append([project_name, p_machine_date[sequencer][project_name], p_machine_lib_kit[sequencer][project_name],
-                                    p_machine_num_sample[sequencer][project_name], '{0:,}'.format(int(p_machine_cluster_pf_dict[sequencer][project_name])),
-                                    '{0:,}'.format(int(p_machine_yield_mb_dict[sequencer][project_name])), p_machine_q30_dict[sequencer][project_name],
-                                    p_machine_mean_dict[sequencer][project_name], p_machine_sequencer[sequencer][project_name]])
-                        sequencer_proj[sequencer] = proj_data
-                        projects_data.append(sequencer_proj)
-                        # create the graphic for q30 quality
-                        theme = 'ocean'
-                        heading = 'Graphics for Q > 30 for machine ' + m_name
-                        sub_caption = 'Sequencer ' + sequencer
-                        x_axis_name = 'Projects'
-                        y_axis_name = 'Q 30 (in %)'
-
-                        data_source = column_graphic_simple (heading, sub_caption, x_axis_name, y_axis_name, theme, p_machine_q30_dict[sequencer])
-                        seq_chart = sequencer + 'q30_chart'
-                        seq_graph = sequencer + 'q30_graph'
-                        q30_machine_seq_graph = FusionCharts("column3d", seq_graph , "500", "350",seq_chart , "json", data_source).render()
-
-                        machine_seq_graphs.append([seq_chart, q30_machine_seq_graph])
-                        # create the graphic for mean quality
-                        theme = 'carbon'
-                        heading = 'Graphics for Mean quality for machine ' + m_name
-                        sub_caption = 'Sequencer ' + sequencer
-                        x_axis_name = 'Projects'
-                        y_axis_name = 'Mean Quality'
-                        data_source = column_graphic_simple (heading, sub_caption, x_axis_name, y_axis_name, theme, p_machine_mean_dict[sequencer])
-                        seq_chart = sequencer + 'mean_q_chart'
-                        seq_graph = sequencer + 'mean_q_graph'
-                        mean_q_machine_seq_graph = FusionCharts("column3d", seq_graph , "500", "350", seq_chart, "json", data_source).render()
-                        machine_seq_graphs.append([seq_chart, mean_q_machine_seq_graph])
-
-                        # create the graphic for yield Mb
-                        theme = 'zune'
-                        heading = 'Graphics for Yield Mb for machine ' + m_name
-                        sub_caption = 'Sequencer ' + sequencer
-                        x_axis_name = 'Projects'
-                        y_axis_name = 'Yield Mb'
-                        data_source = column_graphic_simple (heading, sub_caption, x_axis_name, y_axis_name, theme, p_machine_yield_mb_dict[sequencer])
-                        seq_chart = sequencer + 'yield_mb_chart'
-                        seq_graph = sequencer + 'yield_mb_graph'
-                        yield_mb_machine_graph = FusionCharts("column3d", seq_graph , "500", "350", seq_chart, "json", data_source).render()
-                        machine_seq_graphs.append([seq_chart, yield_mb_machine_graph])
-                        # create the graphic for cluster Pf
-                        theme = 'ocean'
-                        heading = 'Graphics for Cluster Pf for machine ' + m_name
-                        sub_caption = 'Sequencer ' + sequencer
-                        x_axis_name = 'Projects'
-                        y_axis_name = 'Cluster Pf'
-                        data_source = column_graphic_simple (heading, sub_caption, x_axis_name, y_axis_name, theme, p_machine_cluster_pf_dict[sequencer])
-                        seq_chart = sequencer + 'cluster_pf_chart'
-                        seq_graph = sequencer + 'cluster_pf_graph'
-                        cluster_pf_machine_graph = FusionCharts("column3d", seq_graph , "500", "350", seq_chart, "json", data_source).render()
-                        machine_seq_graphs.append([seq_chart, cluster_pf_machine_graph])
-
-
-                        machine_graphs.append(machine_seq_graphs)
-
-                    sequencer_statistics ['machine_graph'] = machine_graphs
-                    sequencer_statistics ['machine_name'] = m_name
-                    sequencer_statistics['projects_data'] = projects_data
-
-
-                    #collecting data for comparation graphics
-
-                    # Calculating the mean for all projects performed by researcher
-                    comp_q30_dict, comp_mean_q_dict = {} , {}
-                    comp_yield_mb_dict, comp_cluster_pf_dict = {} , {}
-
-                    q30_val = p_machine_q30_dict.values()
-                    mean_q_val = p_machine_mean_dict.values()
-                    yield_mb_val = p_machine_yield_mb_dict.values()
-                    cluster_pf = p_machine_cluster_pf_dict.values()
-
-                    for sequencer in projects_name_dict.keys() :
-                        #sequencer_proj = {}
-                        #proj_data =[]
-                        comp_q30_dict[sequencer] , comp_mean_q_dict [sequencer]= {} , {}
-                        comp_yield_mb_dict[sequencer], comp_cluster_pf_dict [sequencer] = {}, {}
-                        comp_q30_dict [sequencer][m_name] = format(statistics.mean( [float(x) for x in list(p_machine_q30_dict[sequencer].values())]),'.2f')
-                        comp_mean_q_dict[sequencer] [m_name] = format(statistics.mean( [float(x) for x in list(p_machine_mean_dict[sequencer].values())]),'.2f')
-
-                        comp_yield_mb_dict[sequencer] [m_name] = sum(list(p_machine_yield_mb_dict[sequencer].values()))
-                        comp_cluster_pf_dict[sequencer] [m_name] = sum(list(p_machine_cluster_pf_dict[sequencer].values()))
-
-                    total_q_30_list, total_mean_q_list = [] , []
-                    total_yield_mb_list, total_cluster_pf_list = [] , []
-                    total_lanes_summary = {}
-
-                    for sequencer in projects_name_dict.keys() :
-                        runs_sequencer = RunProcess.objects.filter(sequencerModel__machineName__exact = sequencer)
-                        run_sequencer_id_list = []
-                        for run in runs_sequencer :
-                            run_sequencer_id_list.append(run.pk)
-
-                        if StatsLaneSummary.objects.filter(runprocess_id__in  = run_sequencer_id_list).exclude(defaultAll__isnull = False).exclude(project_id__in = projects_id_list[sequencer]).exists():
-                            total_lanes_summary[sequencer] = StatsLaneSummary.objects.filter(runprocess_id__in  = run_sequencer_id_list).exclude(defaultAll__isnull = False).exclude(project_id__in = projects_id_list[sequencer])
-                        else:
-                            total_lanes_summary[sequencer] = ''
-
-
-                    comp_graphs, comp_seq_graphs = [] , []
-                    for sequencer in projects_name_dict.keys() :
-                        for lane_summary in total_lanes_summary[sequencer] :
-                            q_30_value, mean_q_value , yield_mb_value , cluster_pf_value = lane_summary.get_stats_info()
-                            total_q_30_list.append(float(q_30_value))
-                            total_mean_q_list.append(float(mean_q_value))
-                            total_yield_mb_list.append(int(yield_mb_value.replace(',','')))
-                            total_cluster_pf_list.append(int(cluster_pf_value.replace(',','')))
-
-
-                    sequencer_statistics ['comp_graphs'] = comp_graphs
-
-                    # Sequencer graphic utilization
-                    sequencer_used = {}
-                    for sequencer in projects_name_dict.keys() :
-                        sequencer_used[sequencer] = len( projects_name_dict[sequencer])
-
-                    theme = 'ocean'
-                    heading = 'Sequencer utilization for machine ' + m_name
-                    sub_caption = ''
-                    data_source = pie_graphic_standard (heading, sub_caption, theme, sequencer_used)
-                    sequencer_pie_graph = FusionCharts("pie3d", "sequencer_pie_graph" , "500", "400", "sequencer_pie_chart", "json", data_source).render()
-                    sequencer_statistics ['sequencer_pie_graph'] = sequencer_pie_graph
-
-                    return  render(request, 'iSkyLIMS_wetlab/StatsPerMachine.html', {'sequencer_statistics' : sequencer_statistics})
-        """
         return render(
             request,
             "iSkyLIMS_wetlab/StatsPerSequencer.html",
